Moved wallet into seperate folder
